--- a/sgl-kernel/csrc/moe/moe_align_kernel.cu
+++ b/sgl-kernel/csrc/moe/moe_align_kernel.cu
@@ -99,55 +99,61 @@ __global__ void moe_align_block_size_kernel(
 }
 
 template <typename scalar_t>
-__global__ void moe_align_block_size_small_batch_kernel(
+__global__ void moe_align_block_size_small_batch_expert_kernel(
     const scalar_t* __restrict__ topk_ids,
     int32_t* __restrict__ sorted_token_ids,
     int32_t* __restrict__ expert_ids,
     int32_t* __restrict__ total_tokens_post_pad,
     int32_t num_experts,
-    int32_t padded_num_experts,
     int32_t block_size,
     size_t numel) {
+  const size_t tokens_per_thread = CEILDIV(numel, blockDim.x);
+  const size_t start_idx = threadIdx.x * tokens_per_thread;
+
   extern __shared__ int32_t shared_mem[];
-  int32_t* expert_counts = shared_mem;
-  int32_t* expert_cumsum = &expert_counts[padded_num_experts];
+  int32_t* cumsum = shared_mem;
+  int32_t* tokens_cnts = (int32_t*)(shared_mem + num_experts + 1);
 
-  const int tid = threadIdx.x;
-  const int num_threads = blockDim.x;
+  for (int i = 0; i < num_experts; ++i) {
+    tokens_cnts[(threadIdx.x + 1) * num_experts + i] = 0;
+  }
 
-  for (int i = tid; i < padded_num_experts; i += num_threads) {
-    expert_counts[i] = 0;
+  for (int i = start_idx; i < numel && i < start_idx + tokens_per_thread; ++i) {
+    ++tokens_cnts[(threadIdx.x + 1) * num_experts + topk_ids[i]];
   }
 
   __syncthreads();
 
-  for (int i = tid; i < numel; i += num_threads) {
-    int32_t expert_id = topk_ids[i];
-    atomicAdd(&expert_counts[expert_id], 1);
+  if (threadIdx.x < num_experts) {
+    tokens_cnts[threadIdx.x] = 0;
+    for (int i = 1; i <= blockDim.x; ++i) {
+      tokens_cnts[i * num_experts + threadIdx.x] += tokens_cnts[(i - 1) * num_experts + threadIdx.x];
+    }
   }
 
   __syncthreads();
 
-  if (tid == 0) {
-    expert_cumsum[0] = 0;
+  if (threadIdx.x == 0) {
+    cumsum[0] = 0;
     for (int i = 1; i <= num_experts; ++i) {
-      expert_cumsum[i] = expert_cumsum[i - 1] + CEILDIV(expert_counts[i - 1], block_size) * block_size;
+      cumsum[i] = cumsum[i - 1] + CEILDIV(tokens_cnts[blockDim.x * num_experts + i - 1], block_size) * block_size;
     }
-    *total_tokens_post_pad = expert_cumsum[num_experts];
+    *total_tokens_post_pad = static_cast<int32_t>(cumsum[num_experts]);
   }
 
   __syncthreads();
 
-  if (tid < num_experts) {
-    for (int i = expert_cumsum[tid]; i < expert_cumsum[tid + 1]; i += block_size) {
-      expert_ids[i / block_size] = tid;
+  if (threadIdx.x < num_experts) {
+    for (int i = cumsum[threadIdx.x]; i < cumsum[threadIdx.x + 1]; i += block_size) {
+      expert_ids[i / block_size] = threadIdx.x;
     }
   }
 
-  for (int i = tid; i < numel; i += num_threads) {
+  for (int i = start_idx; i < numel && i < start_idx + tokens_per_thread; ++i) {
     int32_t expert_id = topk_ids[i];
-    int32_t rank_post_pad = atomicAdd(&expert_cumsum[expert_id], 1);
+    int32_t rank_post_pad = tokens_cnts[threadIdx.x * num_experts + expert_id] + cumsum[expert_id];
     sorted_token_ids[rank_post_pad] = i;
+    ++tokens_cnts[threadIdx.x * num_experts + expert_id];
   }
 }
 
@@ -170,21 +176,19 @@ void moe_align_block_size(
   threads = ((threads + WARP_SIZE - 1) / WARP_SIZE) * WARP_SIZE;
 
   DISPATCH_INTEGRAL_TYPES(topk_ids.scalar_type(), "moe_align_block_size_kernel", [&] {
-    bool small_batch_mode = topk_ids.numel() < 2048;
+    bool small_batch_expert_mode = (topk_ids.numel() < 1024) && (num_experts <= 64);
 
-    if (small_batch_mode) {
-      threads = max((int)num_experts, WARP_SIZE);
+    if (small_batch_expert_mode) {
+      const int32_t num_thread = max((int32_t)num_experts, WARP_SIZE);
+      const int32_t shared_mem_i32 = ((num_thread + 1) * num_experts + (num_experts + 1)) * sizeof(int32_t);
 
-      size_t shared_mem_size = 2 * padded_num_experts * sizeof(int32_t);
-
-      auto small_batch_kernel = moe_align_block_size_small_batch_kernel<scalar_t>;
-      small_batch_kernel<<<1, threads, shared_mem_size, stream>>>(
+      auto small_batch_expert_kernel = moe_align_block_size_small_batch_expert_kernel<scalar_t>;
+      small_batch_expert_kernel<<<1, num_thread, shared_mem_i32, stream>>>(
           topk_ids.data_ptr<scalar_t>(),
           sorted_token_ids.data_ptr<int32_t>(),
           experts_ids.data_ptr<int32_t>(),
           num_tokens_post_pad.data_ptr<int32_t>(),
           num_experts,
-          padded_num_experts,
           block_size,
           topk_ids.numel());
     } else {