<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

- [x] Fix `moe_align_block_size` kernel benchmark bug.
- [x] Reduce moe_align_block_size_kernel small batch mode ovehead caused by `torch.zeros` .
- [x] After this pr merged and new sgl-kernel version released, I can change the `cumsum_buffer` allocate from `torch.zeros` to `torch.empty` in `fused_moe_triton.py`. For small token mode, I wrote a new simple kernel to handle it, and it's not need to use `torch.zeros` to init the `cumsum` buffer, all things are happed in shm. 

## Acc test

I set `token_cnts_buffer` and `cumsum_buffer` to `torch.empty` in `fused_moe.py`:

```python
token_cnts_buffer = torch.empty(
            (num_experts + 1) * num_experts,
            dtype=torch.int32,
            device=topk_ids.device,
        )
        cumsum_buffer = torch.empty(
            num_experts + 1, dtype=torch.int32, device=topk_ids.device
        )
```

Acc result:

```shell
➜  sglang python3 benchmark/gsm8k/bench_sglang.py --num-questions 2000 --parallel 2000 --num-shots 8 --port 30001

100%|████████████████████████████████████████████████████████████████████████| 1319/1319 [01:25<00:00, 15.36it/s]
Accuracy: 0.952
Invalid: 0.000
Latency: 90.543 s
Output throughput: 1532.037 token/s
```

## Kernel unit-test

![图片](https://github.com/user-attachments/assets/b8cd621e-dd61-41d6-8058-20ecf472327f)


## Benchmark In H200

main branch:

```shell
📊 Running performance benchmark for 8 experts...
moe-align-block-size-performance:
     num_tokens  num_experts  topk        SGL       Triton        VLLM
0           1.0          8.0   1.0  18.975999    67.359999   14.624000
1           1.0          8.0   2.0  19.136000    23.264000   14.607999
2           1.0          8.0   4.0  20.384001    63.519999   14.624000
3           1.0          8.0   8.0  19.424001    62.368002   14.656000
4           1.0         32.0   1.0  20.128001    62.912002   16.640000
5           1.0         32.0   2.0  21.536000    63.263997   16.640000
6           1.0         32.0   4.0  21.632001    69.023997   16.576000
7           1.0         32.0   8.0  20.256000    59.712000   16.640000
8           1.0         64.0   1.0  22.816001    56.912001   20.128001
9           1.0         64.0   2.0  22.816001    28.672000   20.256000
10          1.0         64.0   4.0  22.784000    69.440000   20.223999
11          1.0         64.0   8.0  22.816001    65.024003   20.288000
12          1.0        128.0   1.0  24.224000    64.511999   30.975999
13          1.0        128.0   2.0  23.040000    68.335995   30.944001
14          1.0        128.0   4.0  24.288001    63.167997   30.944001
15          1.0        128.0   8.0  23.135999    63.808002   31.040000
16          1.0        256.0   1.0  26.559999    58.240000   65.471999
17          1.0        256.0   2.0  26.591999    70.528001   65.471999
18          1.0        256.0   4.0  26.815999    61.471999   65.632001
19          1.0        256.0   8.0  27.872000    59.680000   65.664001
20          8.0          8.0   1.0  19.200001    65.568000   14.688000
21          8.0          8.0   2.0  19.455999    60.447998   14.688000
22          8.0          8.0   4.0  20.320000    69.552004   14.544000
23          8.0          8.0   8.0  20.447999    68.736002   14.720000
24          8.0         32.0   1.0  21.600001    60.208000   16.640000
25          8.0         32.0   2.0  21.663999    59.999999   16.608000
26          8.0         32.0   4.0  20.576000    62.399998   16.576000
27          8.0         32.0   8.0  21.632001    60.095999   16.672000
28          8.0         64.0   1.0  21.376001    27.264001   20.223999
29          8.0         64.0   2.0  21.376001    60.112000   20.320000
30          8.0         64.0   4.0  21.344000    62.816001   20.223999
31          8.0         64.0   8.0  22.911999    27.360000   20.320000
32          8.0        128.0   1.0  24.288001    57.535999   30.975999
33          8.0        128.0   2.0  24.256000    57.856001   31.104000
34          8.0        128.0   4.0  23.135999    64.544000   31.104000
35          8.0        128.0   8.0  24.224000    57.392001   31.136001
36          8.0        256.0   1.0  27.872000    69.199994   65.632001
37          8.0        256.0   2.0  26.591999    63.040003   65.632001
38          8.0        256.0   4.0  27.872000    55.744000   65.664001
39          8.0        256.0   8.0  27.872000    69.888003   65.696001
40         16.0          8.0   1.0  20.352000    58.944002   14.592000
41         16.0          8.0   2.0  20.352000    53.440001   14.560000
42         16.0          8.0   4.0  19.200001    55.904001   14.720000
43         16.0          8.0   8.0  20.416001    71.744002   14.944000
44         16.0         32.0   1.0  21.600001    65.295994   16.640000
45         16.0         32.0   2.0  20.064000    66.111997   16.576000
46         16.0         32.0   4.0  20.160001    58.543999   16.704001
47         16.0         32.0   8.0  21.600001   156.992003   16.960001
48         16.0         64.0   1.0  21.376001    63.135996   20.288000
49         16.0         64.0   2.0  22.752000    61.792001   20.256000
50         16.0         64.0   4.0  21.344000    63.792005   20.320000
51         16.0         64.0   8.0  21.824000    57.535999   20.416001
52         16.0        128.0   1.0  23.040000    70.175998   31.120000
53         16.0        128.0   2.0  24.224000    63.936003   31.040000
54         16.0        128.0   4.0  23.135999    68.640001   31.168001
55         16.0        128.0   8.0  23.232000    57.119999   31.231999
56         16.0        256.0   1.0  27.904000    66.239998   65.696001
57         16.0        256.0   2.0  26.976001    68.191998   65.664001
58         16.0        256.0   4.0  26.815999    57.280000   65.664001
59         16.0        256.0   8.0  27.968001    62.688001   65.888003
60         32.0          8.0   1.0  20.320000    68.127997   14.560000
61         32.0          8.0   2.0  20.320000    63.167997   14.720000
62         32.0          8.0   4.0  20.352000    56.063998   14.976000
63         32.0          8.0   8.0  19.040000    68.223998   15.552000
64         32.0         32.0   1.0  20.064000    61.280001   16.608000
65         32.0         32.0   2.0  21.663999    59.776001   16.704001
66         32.0         32.0   4.0  21.504000    55.968001   16.992001
67         32.0         32.0   8.0  20.160001    64.000003   17.344000
68         32.0         64.0   1.0  21.824000    61.951999   20.223999
69         32.0         64.0   2.0  21.856001    60.896002   20.320000
70         32.0         64.0   4.0  21.856001    54.048002   20.416001
71         32.0         64.0   8.0  21.632001    72.704002   20.864001
72         32.0        128.0   1.0  23.072001    65.888003   31.136001
73         32.0        128.0   2.0  22.944000    60.063999   31.104000
74         32.0        128.0   4.0  23.072001    70.335999   31.199999
75         32.0        128.0   8.0  23.391999    60.288001   31.328000
76         32.0        256.0   1.0  27.872000    67.456000   65.728001
77         32.0        256.0   2.0  27.775999    67.071997   65.664001
78         32.0        256.0   4.0  27.872000    61.567999   65.792002
79         32.0        256.0   8.0  26.815999    69.023997   65.920003
80         64.0          8.0   1.0  20.479999    58.272000   14.720000
81         64.0          8.0   2.0  20.576000    65.087996   14.944000
82         64.0          8.0   4.0  20.064000    65.375999   15.552000
83         64.0          8.0   8.0  19.424001    72.959997   16.960001
84         64.0         32.0   1.0  21.600001    68.624005   16.736001
85         64.0         32.0   2.0  21.695999    67.376003   16.928000
86         64.0         32.0   4.0  21.504000    64.095996   17.344000
87         64.0         32.0   8.0  21.856001    64.159997   19.200001
88         64.0         64.0   1.0  21.600001    70.367999   20.320000
89         64.0         64.0   2.0  22.848001    68.847999   20.447999
90         64.0         64.0   4.0  21.952000    62.912002   20.800000
91         64.0         64.0   8.0  22.368001    65.728001   21.056000
92         64.0        128.0   1.0  23.167999    75.471997   31.136001
93         64.0        128.0   2.0  23.232000    32.575998   31.296000
94         64.0        128.0   4.0  23.391999    67.727998   31.296000
95         64.0        128.0   8.0  24.383999    60.672000   31.968001
96         64.0        256.0   1.0  26.591999    74.047998   65.728001
97         64.0        256.0   2.0  27.872000    66.463999   65.888003
98         64.0        256.0   4.0  28.063999    64.095996   65.888003
99         64.0        256.0   8.0  26.591999    69.824003   66.111997
100       128.0          8.0   1.0  19.168001    68.896003   14.944000
101       128.0          8.0   2.0  20.096000    60.768001   15.568000
102       128.0          8.0   4.0  20.608000    67.039996   16.928000
103       128.0          8.0   8.0  20.191999    69.632001   20.864001
104       128.0         32.0   1.0  21.536000    67.135997   16.960001
105       128.0         32.0   2.0  21.504000    68.000004   17.344000
106       128.0         32.0   4.0  22.016000    64.223997   19.231999
107       128.0         32.0   8.0  21.888001    68.672001   23.296000
108       128.0         64.0   1.0  22.879999    73.504001   20.416001
109       128.0         64.0   2.0  21.663999    59.712000   20.832000
110       128.0         64.0   4.0  21.728000    66.624001   21.056000
111       128.0         64.0   8.0  21.280000    68.832003   22.560000
112       128.0        128.0   1.0  23.296000    60.095999   31.231999
113       128.0        128.0   2.0  22.911999    70.015997   31.296000
114       128.0        128.0   4.0  24.416000    70.464000   32.032002
115       128.0        128.0   8.0  23.328001    71.327999   32.960001
116       128.0        256.0   1.0  27.039999   216.959998   65.888003
117       128.0        256.0   2.0  28.031999    65.743998   65.856002
118       128.0        256.0   4.0  27.872000    69.632001   66.207998
119       128.0        256.0   8.0  27.008001    68.688005   66.463999
120       256.0          8.0   1.0  20.128001    58.143999   15.536000
121       256.0          8.0   2.0  20.671999    63.231997   16.960001
122       256.0          8.0   4.0  20.223999    65.600000   20.896001
123       256.0          8.0   8.0  21.695999    71.744002   28.287999
124       256.0         32.0   1.0  21.632001    59.103999   17.344000
125       256.0         32.0   2.0  22.016000    29.056000   19.231999
126       256.0         32.0   4.0  20.927999    68.159997   23.264000
127       256.0         32.0   8.0  22.336001    67.520000   30.432001
128       256.0         64.0   1.0  21.632001    29.888000   20.864001
129       256.0         64.0   2.0  22.592001    58.591999   21.088000
130       256.0         64.0   4.0  21.728000    58.623999   22.399999
131       256.0         64.0   8.0  23.744000    66.367999   27.680000
132       256.0        128.0   1.0  23.391999    65.600000   31.296000
133       256.0        128.0   2.0  23.135999    63.023999   32.000002
134       256.0        128.0   4.0  24.704000    68.031996   32.960001
135       256.0        128.0   8.0  23.744000    61.919998   36.031999
136       256.0        256.0   1.0  28.224001    68.768002   65.920003
137       256.0        256.0   2.0  26.591999    65.952003   66.192001
138       256.0        256.0   4.0  27.807999    52.064002   66.479996
139       256.0        256.0   8.0  28.896000    58.784001   68.191998
140       512.0          8.0   1.0  19.231999    69.408000   16.960001
141       512.0          8.0   2.0  19.455999    58.432002   20.800000
142       512.0          8.0   4.0  20.512000    69.215998   28.320000
143       512.0          8.0   8.0  21.056000   114.047997   42.080000
144       512.0         32.0   1.0  21.919999    59.119999   19.200001
145       512.0         32.0   2.0  20.992000    59.071999   23.264000
146       512.0         32.0   4.0  21.536000    58.079999   30.400001
147       512.0         32.0   8.0  23.072001    59.840001   44.512000
148       512.0         64.0   1.0  22.464000    58.304001   21.136001
149       512.0         64.0   2.0  21.663999    32.256000   22.528000
150       512.0         64.0   4.0  23.424000    50.687999   28.112000
151       512.0         64.0   8.0  23.808001    63.519999   41.184001
152       512.0        128.0   1.0  23.040000    59.840001   31.968001
153       512.0        128.0   2.0  23.296000    65.888003   32.960001
154       512.0        128.0   4.0  23.647999    69.327995   35.872001
155       512.0        128.0   8.0  24.192000    69.343999   43.040000
156       512.0        256.0   1.0  26.848000    69.087997   66.111997
157       512.0        256.0   2.0  26.912000    70.319995   66.399999
158       512.0        256.0   4.0  28.928000    56.832001   68.159997
159       512.0        256.0   8.0  28.928000    65.952003   71.392000
160      1024.0          8.0   1.0  20.223999    73.151998   20.896001
161      1024.0          8.0   2.0  21.632001    69.087997   28.352000
162      1024.0          8.0   4.0  22.048000   113.696001   42.032000
163      1024.0          8.0   8.0  25.504000   206.880003   71.039997
164      1024.0         32.0   1.0  21.952000    63.648000   23.264000
165      1024.0         32.0   2.0  21.504000    66.512004   30.400001
166      1024.0         32.0   4.0  22.080000    63.167997   44.480000
167      1024.0         32.0   8.0  24.752000   166.495994   78.368001
168      1024.0         64.0   1.0  21.183999    69.728002   22.464000
169      1024.0         64.0   2.0  23.488000    57.376001   27.664000
170      1024.0         64.0   4.0  22.879999    57.744000   40.959999
171      1024.0         64.0   8.0  26.303999    70.464000   65.183997
172      1024.0        128.0   1.0  23.776000    68.095997   32.864001
173      1024.0        128.0   2.0  23.680000    62.112000   35.824001
174      1024.0        128.0   4.0  24.544001    57.599999   43.040000
175      1024.0        128.0   8.0  27.616000    67.071997   55.135999
176      1024.0        256.0   1.0  26.912000    55.408001   66.463999
177      1024.0        256.0   2.0  27.456000    68.448000   68.191998
178      1024.0        256.0   4.0  27.968001    60.768001   71.295999
179      1024.0        256.0   8.0  30.304000    62.080000   81.887998
180      2048.0          8.0   1.0  20.256000    69.151998   28.320000
181      2048.0          8.0   2.0  20.927999   113.920003   42.048000
182      2048.0          8.0   4.0  25.504000   207.039997   71.071997
183      2048.0          8.0   8.0  32.224000   378.208011  127.519995
184      2048.0         32.0   1.0  22.464000    63.519999   30.432001
185      2048.0         32.0   2.0  23.040000    62.272001   44.447999
186      2048.0         32.0   4.0  23.488000    72.512001   79.392001
187      2048.0         32.0   8.0  29.152000   119.616002  146.559998
188      2048.0         64.0   1.0  23.712000    59.424002   27.295999
189      2048.0         64.0   2.0  22.976000    64.640000   40.927999
190      2048.0         64.0   4.0  25.184000    57.472002   65.183997
191      2048.0         64.0   8.0  31.936001    75.903997  110.271998
192      2048.0        128.0   1.0  23.744000    65.952003   35.744000
193      2048.0        128.0   2.0  25.472000    53.984001   43.072000
194      2048.0        128.0   4.0  26.368000    58.848001   55.135999
195      2048.0        128.0   8.0  32.191999    62.080000   79.328001
196      2048.0        256.0   1.0  28.896000    48.416000   68.223998
197      2048.0        256.0   2.0  27.968001    66.111997   71.359999
198      2048.0        256.0   4.0  31.199999    59.872001   81.791997
199      2048.0        256.0   8.0  37.216000    66.111997   98.272003
200      4096.0          8.0   1.0  20.927999   113.920003   42.048000
201      4096.0          8.0   2.0  25.504000   206.624001   71.039997
202      4096.0          8.0   4.0  31.136001   378.224015  127.424002
203      4096.0          8.0   8.0  43.839999   729.439974  260.639995
204      4096.0         32.0   1.0  21.952000    60.112000   44.544000
205      4096.0         32.0   2.0  23.520000    72.255999   78.847997
206      4096.0         32.0   4.0  29.152000   119.616002  147.872001
207      4096.0         32.0   8.0  46.144001   211.807996  294.223994
208      4096.0         64.0   1.0  22.848001    59.551999   40.991999
209      4096.0         64.0   2.0  25.376000    58.688000   65.279998
210      4096.0         64.0   4.0  30.975999    75.935997  110.207997
211      4096.0         64.0   8.0  49.504001   122.432001  200.703993
212      4096.0        128.0   1.0  24.544001    71.392000   43.040000
213      4096.0        128.0   2.0  27.488001    48.799999   55.232000
214      4096.0        128.0   4.0  32.224000    60.192000   79.328001
215      4096.0        128.0   8.0  49.311999    83.839998  129.695997
216      4096.0        256.0   1.0  28.063999    60.192000   71.263999
217      4096.0        256.0   2.0  30.272000    67.744002   81.791997
218      4096.0        256.0   4.0  36.928002    68.432003   98.240003
219      4096.0        256.0   8.0  49.472000    78.560002  132.175997
220      8192.0          8.0   1.0  26.880000   206.944004   71.039997
221      8192.0          8.0   2.0  32.095999   378.127992  127.568007
222      8192.0          8.0   4.0  43.968000   730.080009  260.320008
223      8192.0          8.0   8.0  69.408000  1435.008049  507.327974
224      8192.0         32.0   1.0  23.520000    72.832003   78.656003
225      8192.0         32.0   2.0  29.216001   119.935997  147.072002
226      8192.0         32.0   4.0  47.072001   212.495998  291.584015
227      8192.0         32.0   8.0  71.071997   386.880010  612.031996
228      8192.0         64.0   1.0  26.591999    53.760000   65.311998
229      8192.0         64.0   2.0  30.944001    76.063998  110.111997
230      8192.0         64.0   4.0  49.791999   122.528002  200.800002
231      8192.0         64.0   8.0  77.632003   215.792000  382.863998
232      8192.0        128.0   1.0  26.400000    60.031999   55.167999
233      8192.0        128.0   2.0  32.448001    60.640000   79.392001
234      8192.0        128.0   4.0  49.472000    83.871998  129.728004
235      8192.0        128.0   8.0  73.792003   131.487995  228.960007
236      8192.0        256.0   1.0  30.368000    62.928006   81.823997
237      8192.0        256.0   2.0  37.248001    65.792002   98.112002
238      8192.0        256.0   4.0  50.592002    78.688003  132.128000
239      8192.0        256.0   8.0  75.328000   105.439998  448.480010
```


pr:

```shell
     num_tokens  num_experts  topk        SGL       Triton        VLLM
0           1.0          8.0   1.0  13.680000    52.671999   13.792000
1           1.0          8.0   2.0  13.664000    53.215999   13.824000
2           1.0          8.0   4.0  13.696000    53.824000   13.856000
3           1.0          8.0   8.0  13.728000    52.384000   13.856000
4           1.0         32.0   1.0  16.416000    55.488002   16.287999
5           1.0         32.0   2.0  16.448000    54.432001   16.416000
6           1.0         32.0   4.0  16.448000    53.024001   16.352000
7           1.0         32.0   8.0  16.448000    53.584002   16.384000
8           1.0         64.0   1.0  19.680001    49.088001   19.520000
9           1.0         64.0   2.0  19.680001    50.496001   19.392001
10          1.0         64.0   4.0  19.711999    50.976001   19.487999
…