@BBuf 2025-04-06
> @BBuf Can you successfully use the tuning script after this change?

It's ok. Tuning bug has been fixed in [pr 4918](https://github.com/sgl-project/sglang/pull/4918)

![图片](https://github.com/user-attachments/assets/ee919c7f-a446-4c12-a6e5-d5f554a8fc60)