@BBuf 2025-04-06
@zhyncs @merrymercy  Now, I have fixed all the bug and performace bug in sgl_kernel `moe_align_block_size` kernel. It can be seen on the H200 benchmark that the performance in all scenarios is now on par with or better than VLLM. As the number of tokens and the number of experts increase, the speedup further improves. Once the CI passed, we can merge it and I can change the cumsum_buffer allocate from torch.zeros to torch.empty in fused_moe_triton.py when new sgl-kernel version released.