@BBuf 2025-04-10
```shell
    num_tokens  num_experts  topk         SGL        Triton         VLLM
0      16384.0          8.0   1.0   27.744001    377.983987   126.864001
1      16384.0          8.0   2.0   36.991999    733.215988   259.519994
2      16384.0          8.0   4.0   55.583999   1440.768003   508.512020
3      16384.0          8.0   8.0   92.384003   2859.999895  1001.136065
4      16384.0         32.0   1.0   28.767999    119.584002   146.592006
5      16384.0         32.0   2.0   38.943999    212.543994   295.103997
6      16384.0         32.0   4.0   58.304001    385.760009   608.287990
7      16384.0         32.0   8.0   98.463997    745.248020  1175.904036
8      16384.0         64.0   1.0   30.944001     76.448001   109.856002
9      16384.0         64.0   2.0   41.760001    123.744003   200.320005
10     16384.0         64.0   4.0   63.584000    216.831997   382.048011
11     16384.0         64.0   8.0  107.199997    390.175998   755.392015
12     16384.0        128.0   1.0   30.912001     59.136000    79.296000
13     16384.0        128.0   2.0   40.383998     83.360001   129.728004
14     16384.0        128.0   4.0   58.559999    129.951999   228.271991
15     16384.0        128.0   8.0   96.032001    223.072007   426.912010
16     16384.0        256.0   1.0   35.392001     67.327999    98.463997
17     16384.0        256.0   2.0   41.439999     80.192000   131.871998
18     16384.0        256.0   4.0   57.760000    102.816001   448.000014
19     16384.0        256.0   8.0   90.559997    152.288005   762.592018
20     32768.0          8.0   1.0   37.216000    738.048017   259.552002
21     32768.0          8.0   2.0   55.536002   1451.167941   507.488012
22     32768.0          8.0   4.0   92.896000   2860.960007  1002.847910
23     32768.0          8.0   8.0  169.760004   5724.895954  1990.640044
24     32768.0         32.0   1.0   39.039999    212.543994   293.888003
25     32768.0         32.0   2.0   58.688000    386.511981   610.943973
26     32768.0         32.0   4.0   98.623998    745.696008  1178.751945
27     32768.0         32.0   8.0  180.319995   1468.608022  2393.248081
28     32768.0         64.0   1.0   42.048000    123.039998   200.703993
29     32768.0         64.0   2.0   63.231997    216.352001   382.432014
30     32768.0         64.0   4.0  107.648000    390.464008   751.215994
31     32768.0         64.0   8.0  199.072003    752.991974  1491.999984
32     32768.0        128.0   1.0   40.320002     83.376005   129.567996
33     32768.0        128.0   2.0   58.816001    130.991995   228.863999
34     32768.0        128.0   4.0   96.064001    223.744005   427.583992
35     32768.0        128.0   8.0  172.992006    399.744004   844.015956
36     32768.0        256.0   1.0   41.471999     78.783996   131.935999
37     32768.0        256.0   2.0   57.696000    103.200004   449.088007
38     32768.0        256.0   4.0   90.655997    152.383998   763.616025
39     32768.0        256.0   8.0  159.904003    249.952003  1397.055984
40     65536.0          8.0   1.0   55.520002   1451.712012   507.856011
41     65536.0          8.0   2.0   92.799999   2862.272024  1002.287984
42     65536.0          8.0   4.0  168.799996   5728.703976  1989.536047
43     65536.0          8.0   8.0  315.903991  11395.999908  3962.704182
44     65536.0         32.0   1.0   58.816001    386.335999   609.535992
45     65536.0         32.0   2.0   98.495997    746.240020  1169.664025
46     65536.0         32.0   4.0  181.088001   1468.719959  2406.896114
47     65536.0         32.0   8.0  336.800009   2901.887894  4866.208076
48     65536.0         64.0   1.0   63.231997    216.639996   382.016003
49     65536.0         64.0   2.0  107.327998    390.560001   752.560019
50     65536.0         64.0   4.0  198.080003    752.416015  1490.944028
51     65536.0         64.0   8.0  372.415990   1472.352028  2908.512115
52     65536.0        128.0   1.0   58.975998    130.559996   228.720009
53     65536.0        128.0   2.0   96.064001    223.583996   428.351998
54     65536.0        128.0   4.0  174.464002    399.792016   842.736006
55     65536.0        128.0   8.0  322.880000    761.391997  1638.512015
56     65536.0        256.0   1.0   57.728000    103.391998   447.551996
57     65536.0        256.0   2.0   90.767995    152.160004   763.599992
58     65536.0        256.0   4.0  158.672005    249.952003  1397.632003
59     65536.0        256.0   8.0  289.600015    429.791987  2670.127869
60    131072.0          8.0   1.0   92.735998   2864.223957  1001.824021
61    131072.0          8.0   2.0  168.768004   5724.736214  1993.952036
62    131072.0          8.0   4.0  315.744013  11396.672249  3963.071823
63    131072.0          8.0   8.0  606.112003  22729.358673  7922.160149
64    131072.0         32.0   1.0   98.527998    746.016026  1176.031947
65    131072.0         32.0   2.0  181.024000   1468.736053  2392.623901
66    131072.0         32.0   4.0  336.479992   2903.136015  4890.624046
67    131072.0         32.0   8.0  647.647977   5768.127918  9744.992256
68    131072.0         64.0   1.0  107.423998    390.464008   752.864003
69    131072.0         64.0   2.0  198.783994    752.255976  1490.175962
70    131072.0         64.0   4.0  372.447997   1472.336054  2910.624027
71    131072.0         64.0   8.0  717.664003   2899.647951  5952.688217
72    131072.0        128.0   1.0   96.256003    224.127993   427.648008
73    131072.0        128.0   2.0  173.439994    399.776012   843.312025
74    131072.0        128.0   4.0  322.784007    761.744022  1638.623953
75    131072.0        128.0   8.0  617.088020   1481.727958  3243.520021
76    131072.0        256.0   1.0   90.751998    152.416006   763.360023
77    131072.0        256.0   2.0  159.040004    249.727994  1398.591995
78    131072.0        256.0   4.0  289.503992    429.536015  2671.008110
79    131072.0        256.0   8.0  548.527956    797.551990  5192.447662
```

Now the kernel outperforms Triton and VLLM under any conditions of tokens, num_experts, and topk. Previously, it was slower than Triton when the token count was large, such as >=65536, because a counting loop had not been changed to a strided loop, resulting in non-contiguous memory access. cc @yiakwy-xpu-ml-framework-team